fix: BigInt COUNT cast to int

+++ app/api/admin/clientes/route.ts
@@ -7,7 +7,7 @@ export async function GET(req: Request) {
   try {
     const empresa_id = await requireEmpresaId()
     const clientes = await prisma.$queryRaw<any[]>(Prisma.sql`
-      SELECT c.*, COUNT(DISTINCT p.id) AS num_propiedades
+      SELECT c.*, COUNT(DISTINCT p.id)::int AS num_propiedades
       FROM clientes c LEFT JOIN propiedades p ON p.cliente_id = c.id
       WHERE c.empresa_id = ${empresa_id}::uuid
       GROUP BY c.id ORDER BY c.nombre
